v0.5 (#24)
* Add experimental grammar file

* Update dependency tslint to v6.1.1

* Add SQL file parsing

* Update dependency pg to v8

* SQL file syntax support

* Pure SQL query parsing and generation

* Add prettier

* Prettify project

* Update example to use SQL syntax

* Better error reporting and queue fixes

* Update dependency ts-jest to v25.4.0

* Update docs

+++ packages/cli/src/generator.test.ts
@@ -1,51 +1,61 @@
-import * as queryModule from "@pgtyped/query";
-import {
-  FieldType, generateInterface, queryToTypeDeclarations,
-} from "./generator";
+import * as queryModule from '@pgtyped/query';
+import { generateInterface, queryToTypeDeclarations } from './generator';
+import { ProcessingMode } from './index';
+import { parseSQLFile } from '@pgtyped/query';
 
-const getTypesMocked = jest
-  .spyOn(queryModule, "getTypes")
-  .mockName("getTypes");
+const getTypesMocked = jest.spyOn(queryModule, 'getTypes').mockName('getTypes');
 
-test("test query to interface", async () => {
-  const query = `
-      delete from users *
-      where name = :userName and id = :userId and note = :userNote returning id, id, name, note as bote;
+test('query-to-interface translation (SQL)', async () => {
+  const queryString = `
+    /* @name DeleteUsers */
+      delete from users * where name = :userName and id = :userId and note = :userNote returning id, id, name, note as bote;
     `;
   const mockTypes = {
     returnTypes: [
       {
-        returnName: "id",
-        columnName: "id",
-        typeName: "uuid",
+        returnName: 'id',
+        columnName: 'id',
+        typeName: 'uuid',
         nullable: false,
       },
       {
-        returnName: "name",
-        columnName: "name",
-        typeName: "text",
+        returnName: 'name',
+        columnName: 'name',
+        typeName: 'text',
         nullable: false,
       },
       {
-        returnName: "bote",
-        columnName: "note",
-        typeName: "text",
+        returnName: 'bote',
+        columnName: 'note',
+        typeName: 'text',
         nullable: true,
       },
     ],
     paramMetadata: {
-      params: ["uuid", "text"],
+      params: ['uuid', 'text'],
       mapping: [
-        { name: "id", type: queryModule.ParamType.Scalar, assignedIndex: 1 },
-        { name: "userName", type: queryModule.ParamType.Scalar, assignedIndex: 2 },
+        {
+          name: 'id',
+          type: queryModule.ParamTransform.Scalar,
+          assignedIndex: 1,
+        },
+        {
+          name: 'userName',
+          type: queryModule.ParamTransform.Scalar,
+          assignedIndex: 2,
+        },
       ],
     },
   };
   getTypesMocked.mockResolvedValue(mockTypes as any);
-  const result = await queryToTypeDeclarations({
-    name: "DeleteUsers",
-    body: query,
-  }, null);
+  const query = parseSQLFile(queryString);
+  const result = await queryToTypeDeclarations(
+    {
+      ast: query.parseTree.queries[0],
+      mode: ProcessingMode.SQL,
+    },
+    null,
+  );
   const expected = `/** 'DeleteUsers' parameters type */
 export interface IDeleteUsersParams {
   id: string | null;
@@ -69,21 +79,97 @@ export interface IDeleteUsersQuery {
   expect(result).toEqual(expected);
 });
 
-test("interface generation", () => {
-  const expected =
-    `export interface User {
+test('query-to-interface translation (TS)', async () => {
+  const query = `
+      delete from users *
+      where name = :userName and id = :userId and note = :userNote returning id, id, name, note as bote;
+    `;
+  const mockTypes = {
+    returnTypes: [
+      {
+        returnName: 'id',
+        columnName: 'id',
+        typeName: 'uuid',
+        nullable: false,
+      },
+      {
+        returnName: 'name',
+        columnName: 'name',
+        typeName: 'text',
+        nullable: false,
+      },
+      {
+        returnName: 'bote',
+        columnName: 'note',
+        typeName: 'text',
+        nullable: true,
+      },
+    ],
+    paramMetadata: {
+      params: ['uuid', 'text'],
+      mapping: [
+        {
+          name: 'id',
+          type: queryModule.ParamTransform.Scalar,
+          assignedIndex: 1,
+        },
+        {
+          name: 'userName',
+          type: queryModule.ParamTransform.Scalar,
+          assignedIndex: 2,
+        },
+      ],
+    },
+  };
+  getTypesMocked.mockResolvedValue(mockTypes as any);
+  const result = await queryToTypeDeclarations(
+    {
+      name: 'DeleteUsers',
+      body: query,
+      mode: ProcessingMode.TS,
+    },
+    null,
+  );
+  const expected = `/** 'DeleteUsers' parameters type */
+export interface IDeleteUsersParams {
+  id: string | null;
+  userName: string | null;
+}
+
+/** 'DeleteUsers' return type */
+export interface IDeleteUsersResult {
+  id: string;
+  name: string;
+  bote: string | null;
+}
+
+/** 'DeleteUsers' query type */
+export interface IDeleteUsersQuery {
+  params: IDeleteUsersParams;
+  result: IDeleteUsersResult;
+}
+
+`;
+  expect(result).toEqual(expected);
+});
+
+test('interface generation', () => {
+  const expected = `export interface User {
   name: string;
   age: number;
 }
 
 `;
-  const fields = [{
-    fieldName: "name",
-    fieldType: "string",
-  }, {
-    fieldName: "age",
-    fieldType: "number",
-  }];
-  const result = generateInterface("User", fields);
+  const fields = [
+    {
+      fieldName: 'name',
+      fieldType: 'string',
+    },
+    {
+      fieldName: 'age',
+      fieldType: 'number',
+    },
+  ];
+  const result = generateInterface('User', fields);
   expect(result).toEqual(expected);
 });
